Extract pad opener module
