+++ js/etherpad_nextcloud-files-main.mjs
@@ -1,2 +1,2 @@
-import{a as _,A as f,o as ye,b as xe,c as Ae,d as re,n as Ne,e as Le,M as ie,f as Se}from"./oc-compat-Cn8-bqPr.chunk.mjs";import{p as $,a as we,b as j,g as Q,n as B,i as K,v as G,f as J,c as Me,r as X,d as Oe,e as Te}from"./urls-BxGmTTaP.chunk.mjs";const q=new Map,Z=async e=>{const n=String(e);if(q.has(n))return q.get(n);const r=_("/apps/"+f+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(n),o=U(r,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve failed.").catch(s=>{throw q.delete(n),s});return q.set(n,o),o},ke=async e=>{const n="path:"+String(e);if(q.has(n))return q.get(n);const r=_("/apps/"+f+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(e),o=U(r,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve by path failed.").catch(s=>{throw q.delete(n),s});return q.set(n,o),o},Y=async e=>{const n=_("/apps/"+f+"/api/v1/pads/sync-status/"+encodeURIComponent(String(e)));return U(n,{method:"GET",headers:{Accept:"application/json"}},"Sync status check failed.")},Ie=async(e,n)=>{let r=_("/apps/"+f+"/api/v1/pads/sync/"+encodeURIComponent(String(e)));return r+="?force=1",U(r,{method:"POST",headers:{Accept:"application/json",requesttoken:ye()}},"Pad sync failed.")},Pe=async(e,n)=>{const r=_("/apps/"+f+"/api/v1/pads/from-url"),o=new URLSearchParams;return o.set("file",e),o.set("padUrl",n),U(r,{method:"POST",headers:be(),body:o.toString()},"Could not import public pad URL.")},Fe=async e=>{const n=_("/apps/"+f+"/api/v1/pads"),r=new URLSearchParams;return r.set("file",e),r.set("accessMode","public"),U(n,{method:"POST",headers:be(),body:r.toString()},"Could not create public pad.")},be=()=>({Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:ye()}),U=async(e,n,r)=>{const o=await fetch(e,{...n,credentials:"same-origin"}),s=await o.json().catch(()=>({}));if(!o.ok)throw new Error(s&&s.message||r);return s},H=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),F=()=>(window.location.pathname||"").includes("/apps/files"),qe=120,_e=900,oe=async(e,n={})=>{if(H()&&e.path)try{const r=Number(e.fileId);F()&&Number.isFinite(r)&&r>0&&(ze(r,e.path,n.resolveOpenDir),await Re()),await He(e.path),window.OCA.Viewer.open({path:e.path,onClose:De});return}catch{}typeof n.fallbackOpen=="function"&&await n.fallbackOpen(e)},Re=()=>new Promise(e=>{window.setTimeout(e,qe)}),He=async e=>{const n=xe(e);n!==""&&(Ae("editor:file:created",n),await Ue())},Ue=()=>new Promise(e=>{window.setTimeout(e,_e)}),ze=(e,n,r)=>{const o=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;!o||typeof o.goToRoute!="function"||o.goToRoute(null,{...o.params||{},view:o.params&&o.params.view?o.params.view:"files",fileid:String(e)},{...o.query||{},dir:Be(n,r),editing:"false",openfile:"true"},!0)},Be=(e,n)=>{if(typeof n=="function")return n(e);const r=String(e||""),o=r.lastIndexOf("/");return o>0?r.substring(0,o):"/"},De=()=>{const e=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;if(!e||typeof e.goToRoute!="function")return;const n={...e.query||{}};delete n.openfile,delete n.editing,e.goToRoute(null,e.params||{},n)},ee="data-epnc-sidebar-sync-panel",ne="data-epnc-sidebar-sync-mount",he=8e3,je=6e4,ge=15e3,$e=8e3,Ve=()=>{const e=document.documentElement,n=document.body,r=(e?.className||"")+" "+(n?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(r))return!0;const o=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(o&&o.matches)},ae=e=>{if(!(e instanceof HTMLElement))return!1;const n=window.getComputedStyle(e);return n.display==="none"||n.visibility==="hidden"?!1:n.position==="fixed"?!0:e.offsetParent!==null},We=e=>{const n=Number(e);return Number.isFinite(n)&&n>0?n:null},Ke=()=>{const e=j();if(e!==null)return e;const n=new URLSearchParams(window.location.search||""),r=Number(n.get("fileid")||"");return Number.isFinite(r)&&r>0?r:null},se=()=>{const e=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const n of e){const r=document.querySelector(n);if(r instanceof HTMLElement)return r}return null},Ge=e=>{if(!(e instanceof HTMLElement))return null;let n=e.querySelector("["+ne+'="1"]');n instanceof HTMLElement||(n=document.createElement("div"),n.className="epnc-sidebar-sync-mount",n.setAttribute(ne,"1"));const r=e.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(r&&r.parentNode instanceof HTMLElement){const o=r.parentNode;return(n.parentNode!==o||n.nextSibling!==r)&&o.insertBefore(n,r),n}return n.parentNode!==e&&e.insertBefore(n,e.firstChild),n},le=e=>{const n=Ke();if(n!==null)return n;if(!(e instanceof HTMLElement))return null;const r=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const s of r){const l=e.querySelector(s);if(!(l instanceof HTMLElement))continue;const v=[l.getAttribute("data-fileid"),l.getAttribute("data-file-id"),l.getAttribute("data-node-id"),l.getAttribute("data-id"),l.dataset&&(l.dataset.fileid||l.dataset.fileId)];for(const w of v){const b=We(w);if(b!==null)return b}}const o=e.querySelectorAll("a[href]");for(const s of o){if(!(s instanceof HTMLAnchorElement))continue;const l=we(s.getAttribute("href"));if(l!==null)return l}return null},Je=e=>e==="synced"?"etherpad-icon-color":Ve()?"etherpad-icon-white":"etherpad-icon-black",Ze=e=>e==="synced"?t(f,"Synchronized"):e==="syncing"?t(f,"Saving..."):e==="error"?t(f,"Sync failed"):e==="unavailable"?t(f,"Sync status unavailable"):e==="loading"?t(f,"Checking sync status..."):t(f,"Sync pending"),te=e=>e==="synced"?"synced":e==="out_of_sync"?"pending":e==="unavailable"?"unavailable":"error",Qe=(e,n)=>e==="synced"?Math.min(je,n+$e):e==="pending"?he:ge,Xe=()=>{let e=null,n=null,r=null,o=0;const s=()=>{r!==null&&(window.clearTimeout(r),r=null)},l=()=>{const d=document.querySelector("["+ee+'="1"]');d&&d.parentNode&&d.parentNode.removeChild(d);const u=document.querySelector("["+ne+'="1"]');u&&u.parentNode&&u.parentNode.removeChild(u),s()},v=(d,u)=>{if(!(d instanceof HTMLElement))return;const c=String(u||"").trim();d.setAttribute("data-public-open-url",c);const g=d.querySelector("[data-epnc-sidebar-open-public-button]");if(!(g instanceof HTMLButtonElement))return;const A=c!=="";g.style.display=A?"inline-flex":"none",g.disabled=!A},w=(d,u,c,g,A=!0)=>{if(!(d instanceof HTMLElement))return;const x=d.querySelector("[data-epnc-sidebar-sync-icon]"),S=d.querySelector("[data-epnc-sidebar-sync-label]"),N=d.querySelector("[data-epnc-sidebar-sync-button]");if(d.setAttribute("data-file-id",String(c)),d.setAttribute("data-sync-state",u),d.setAttribute("data-sync-in-flight",g?"1":"0"),S instanceof HTMLElement&&(S.textContent=Ze(u)),x instanceof HTMLElement){const P=Je(u),M=re(f,P);M!==""&&(x.style.backgroundImage=`url(${M})`)}N instanceof HTMLButtonElement&&(N.disabled=g||!A||c<=0,N.setAttribute("aria-busy",g?"true":"false"))},b=(d,u)=>{if(!(u instanceof HTMLElement))return null;let c=u.querySelector("["+ee+'="1"]');if(!(c instanceof HTMLElement)){c=document.createElement("section"),c.className="epnc-sidebar-sync-panel",c.setAttribute(ee,"1"),c.setAttribute("data-file-id",String(d));const g=document.createElement("div");g.className="epnc-sidebar-sync-panel__header";const A=document.createElement("span");A.className="epnc-sidebar-sync-panel__icon",A.setAttribute("data-epnc-sidebar-sync-icon","1"),A.setAttribute("aria-hidden","true");const x=document.createElement("span");x.className="epnc-sidebar-sync-panel__status",x.setAttribute("data-epnc-sidebar-sync-label","1"),g.appendChild(A),g.appendChild(x);const S=document.createElement("button");S.type="button",S.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",S.setAttribute("data-epnc-sidebar-sync-button","1"),S.textContent=t(f,"Pad in Datei speichern"),S.addEventListener("click",async P=>{P.preventDefault(),P.stopPropagation();const M=Number(c.getAttribute("data-file-id")||"");if(!(!Number.isFinite(M)||M<=0)){w(c,"syncing",M,!0,!0);try{await Ie(M,!0);const V=await Y(M),W=te(V&&V.status);w(c,W,M,!1,!0)}catch{w(c,"error",M,!1,!0)}}});const N=document.createElement("button");N.type="button",N.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",N.setAttribute("data-epnc-sidebar-open-public-button","1"),N.textContent=t(f,"Pad in neuem Tab öffnen"),N.style.display="none",N.addEventListener("click",P=>{P.preventDefault(),P.stopPropagation();const M=String(c.getAttribute("data-public-open-url")||"").trim();M!==""&&window.open(M,"_blank","noopener,noreferrer")}),c.appendChild(g),c.appendChild(S),c.appendChild(N)}return c.parentNode!==u&&u.appendChild(c),c},C=(d,u)=>{if(s(),!(d instanceof HTMLElement)||!Number.isFinite(u)||u<=0)return;let c=he;const g=x=>{r=window.setTimeout(()=>{A()},Math.max(1e3,x))},A=async()=>{if(!(d instanceof HTMLElement)||!document.body.contains(d)){s();return}const x=se();if(!(x instanceof HTMLElement)||!ae(x)||!ae(d)){s();return}if(le(x)!==u){s();return}if(document.visibilityState!=="visible"||d.getAttribute("data-sync-in-flight")==="1"){g(c);return}try{const S=await Y(u),N=te(S&&S.status);w(d,N,u,!1,!0),c=Qe(N,c)}catch{w(d,"error",u,!1,!0),c=ge}g(c)};g(c)},k=()=>{n!==null&&(window.clearTimeout(n),n=null)},I=async()=>{const d=++o;if(!F()||$()){l();return}const u=se();if(!(u instanceof HTMLElement)){l();return}const c=le(u);if(!c){l();return}const g=Ge(u);if(!(g instanceof HTMLElement)){l();return}let A;try{A=await Z(c)}catch{l();return}if(d!==o)return;if(!A||A.is_pad!==!0){l();return}const x=b(c,g);if(!(x instanceof HTMLElement)){l();return}const S=A&&typeof A.public_open_url=="string"?A.public_open_url.trim():"";v(x,S),w(x,"loading",c,!1,!0);try{const N=await Y(c);if(d!==o)return;const P=te(N&&N.status);w(x,P,c,!1,!0),C(x,c)}catch{if(d!==o)return;w(x,"error",c,!1,!0),C(x,c)}},h=(d=0)=>{k(),n=window.setTimeout(()=>{n=null,I()},Math.max(0,d))};return{installObserver:()=>{if(e!==null||!F())return;const d=new MutationObserver(u=>{for(const c of u)for(const g of c.addedNodes)if(g instanceof Element&&(g.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||g.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){h(60);return}});d.observe(document.body,{childList:!0,subtree:!0}),e=d},scheduleRefresh:h}},ce=f+"_public_pad",de="icon-filetype-etherpad-nextcloud-pad",ue='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',pe=98,Ye=120,et=500,fe=()=>{const e=Ne();return e!==null&&e>=33},D=e=>(e&&e.message?String(e.message):"").toLowerCase().includes("duplicate"),tt=e=>{if(!e||typeof e!="object")return null;if(e.hasCreatePermission===!1)return!1;if(e.hasCreatePermission===!0)return!0;if(typeof e.canCreate=="boolean")return e.canCreate;const n=Le(),r=[e.permissions,e.permission,e.attributes&&e.attributes.permissions,typeof e.get=="function"?e.get("permissions"):null];for(const o of r){const s=Number(o);if(Number.isFinite(s))return(s&n)===n}return null},nt=(...e)=>{for(const n of e){const r=tt(n);if(r!==null)return r}return!0},rt=()=>{const e=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const r of e)if(!(!r||typeof r!="object")){if(typeof r.addNewFileMenuEntry=="function")return{register:r.addNewFileMenuEntry.bind(r),unregister:typeof r.removeNewFileMenuEntry=="function"?r.removeNewFileMenuEntry.bind(r):null};if(typeof r.getNewFileMenu=="function"){let o=null;try{o=r.getNewFileMenu()}catch{o=null}if(o&&typeof o.registerEntry=="function")return{register:o.registerEntry.bind(o),unregister:typeof o.unregisterEntry=="function"?o.unregisterEntry.bind(o):null}}}const n=window._nc_newfilemenu;return n&&typeof n=="object"&&typeof n.registerEntry=="function"?{register:n.registerEntry.bind(n),unregister:typeof n.unregisterEntry=="function"?n.unregisterEntry.bind(n):null}:null},it=({isFilesAppRoute:e,onCreatePublicPad:n})=>{let r=!1,o=!1,s=!1,l=0;const v=()=>{n()},w=(...h)=>nt(...h),b=()=>({id:ce,displayName:t(f,"Public pad"),...fe()?{iconSvgInline:ue}:{iconClass:de},order:pe,enabled:w,handler:v}),C=()=>({id:ce,displayName:t(f,"Public pad"),templateName:t(f,"Public pad"),iconClass:de,...fe()?{iconSvgInline:ue}:{},fileType:"file",order:pe,actionHandler:v}),k=async()=>{const h=rt();if(!h)return!1;const d=b();try{return h.register(d),r=!0,!0}catch(u){return D(u)?(r=!0,!0):!1}},I=()=>{if(o)return!0;const h=C(),d=()=>{const u=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const c of u)if(!(!c||typeof c!="object")){if(typeof c.addMenuEntry=="function")try{return c.addMenuEntry(h),!0}catch(g){if(D(g))return!0}if(typeof c.registerEntry=="function")try{return c.registerEntry(b()),!0}catch(g){if(D(g))return!0}}return!1};try{return d()?(o=!0,!0):(!s&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(u){if(!(!u||typeof u.addMenuEntry!="function"))try{u.addMenuEntry(h),o=!0}catch(c){D(c)&&(o=!0)}}}),s=!0),o)}catch(u){return D(u)?(o=!0,!0):!1}};return()=>{if(!e()||r||o)return;const h=++l,d=async u=>{h!==l||r||o||await k()||I()||u<Ye&&window.setTimeout(()=>{d(u+1)},et)};d(0)}},ot=e=>{try{const n=new URL(e,window.location.origin),r=decodeURIComponent(n.pathname||"").match(/\/p\/([^/]+)$/);if(r&&r[1]){const o=r[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(o!=="")return o+".pad"}}catch{}return"Imported pad.pad"},ve=e=>{const n=document.createElement("div");n.style.position="fixed",n.style.inset="0",n.style.background="rgba(0, 0, 0, 0.45)",n.style.display="flex",n.style.alignItems="center",n.style.justifyContent="center",n.style.zIndex="20000";const r=document.createElement("div");r.style.position="relative",r.style.background="var(--color-main-background, #fff)",r.style.borderRadius="10px",r.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",r.style.padding="18px",r.style.width="min(460px, calc(100vw - 24px))";const o=document.createElement("button");o.type="button",o.setAttribute("aria-label",t(f,"Close")),o.textContent="×",o.style.position="absolute",o.style.top="8px",o.style.right="10px",o.style.border="none",o.style.background="transparent",o.style.fontSize="22px",o.style.cursor="pointer",o.style.lineHeight="1";const s=document.createElement("h3");return s.textContent=e,s.style.margin="0 26px 10px 0",s.style.fontSize="18px",r.appendChild(o),r.appendChild(s),n.appendChild(r),document.body.appendChild(n),{overlay:n,dialog:r,closeButton:o}},at=()=>new Promise(e=>{const{overlay:n,dialog:r,closeButton:o}=ve(t(f,"Internal public pad")),s=document.createElement("label");s.textContent=t(f,"File name"),s.style.display="block",s.style.marginBottom="6px";const l=document.createElement("input");l.type="text",l.value=t(f,"Public pad")+".pad",l.style.width="100%",l.style.boxSizing="border-box",l.style.marginBottom="12px";const v=document.createElement("p");v.style.color="var(--color-error, #c62828)",v.style.margin="0 0 12px 0",v.style.minHeight="20px";const w=document.createElement("button");w.type="button",w.className="primary",w.textContent=t(f,"Create");const b=C=>{n.remove(),e(C)};o.addEventListener("click",()=>b(null)),n.addEventListener("click",C=>{C.target===n&&b(null)}),l.addEventListener("keydown",C=>{C.key==="Enter"&&(C.preventDefault(),w.click())}),w.addEventListener("click",()=>{const C=l.value.trim();if(C===""){v.textContent=t(f,"File name is required.");return}b(C)}),r.appendChild(s),r.appendChild(l),r.appendChild(v),r.appendChild(w),l.focus(),l.select()}),st=()=>new Promise(e=>{const{overlay:n,dialog:r,closeButton:o}=ve(t(f,"External public pad (URL)")),s=document.createElement("label");s.textContent=t(f,"External pad URL"),s.style.display="block",s.style.marginBottom="6px";const l=document.createElement("input");l.type="url",l.value="https://",l.placeholder="https://",l.style.width="100%",l.style.boxSizing="border-box",l.style.marginBottom="12px";const v=document.createElement("label");v.textContent=t(f,"File name"),v.style.display="block",v.style.marginBottom="6px";const w=document.createElement("input");w.type="text",w.value="Imported pad.pad",w.style.width="100%",w.style.boxSizing="border-box",w.style.marginBottom="12px";const b=document.createElement("p");b.style.color="var(--color-error, #c62828)",b.style.margin="0 0 12px 0",b.style.minHeight="20px";const C=document.createElement("button");C.type="button",C.className="primary",C.textContent=t(f,"Create");const k=h=>{n.remove(),e(h)};o.addEventListener("click",()=>k(null)),n.addEventListener("click",h=>{h.target===n&&k(null)}),l.addEventListener("blur",()=>{const h=l.value.trim();h.startsWith("http")&&(w.value=ot(h))});const I=()=>{const h=l.value.trim(),d=w.value.trim();if(h===""){b.textContent=t(f,"External pad URL is required.");return}if(d===""){b.textContent=t(f,"File name is required.");return}k({padUrl:h,name:d})};l.addEventListener("keydown",h=>{h.key==="Enter"&&(h.preventDefault(),I())}),w.addEventListener("keydown",h=>{h.key==="Enter"&&(h.preventDefault(),I())}),C.addEventListener("click",I),r.appendChild(s),r.appendChild(l),r.appendChild(v),r.appendChild(w),r.appendChild(b),r.appendChild(C),l.focus(),l.select()}),lt=()=>new Promise(e=>{const n=document.createElement("div");n.style.position="fixed",n.style.inset="0",n.style.background="rgba(0, 0, 0, 0.45)",n.style.display="flex",n.style.alignItems="center",n.style.justifyContent="center",n.style.zIndex="20000";const r=document.createElement("div");r.style.position="relative",r.style.background="var(--color-main-background, #fff)",r.style.borderRadius="10px",r.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",r.style.padding="18px",r.style.width="min(420px, calc(100vw - 24px))";const o=document.createElement("h3");o.textContent=t(f,"Create public pad"),o.style.margin="0 26px 10px 0",o.style.fontSize="18px";const s=document.createElement("button");s.type="button",s.setAttribute("aria-label",t(f,"Close")),s.textContent="×",s.style.position="absolute",s.style.top="8px",s.style.right="10px",s.style.border="none",s.style.background="transparent",s.style.fontSize="22px",s.style.cursor="pointer",s.style.lineHeight="1";const l=document.createElement("p");l.textContent=t(f,"Choose how you want to create the public pad."),l.style.margin="0 0 14px 0",l.style.opacity="0.8";const v=document.createElement("div");v.style.display="grid",v.style.gap="8px";const w=document.createElement("button");w.type="button",w.className="primary",w.textContent=t(f,"Internal public pad");const b=document.createElement("button");b.type="button",b.textContent=t(f,"External public pad (URL)");const C=re(f,"etherpad-icon-color");C!==""&&(b.style.backgroundImage=`url(${C})`),b.style.backgroundRepeat="no-repeat",b.style.backgroundPosition="12px center",b.style.backgroundSize="16px 16px",b.style.paddingLeft="34px";const k=I=>{n.remove(),e(I)};s.addEventListener("click",()=>k(null)),w.addEventListener("click",()=>k("internal")),b.addEventListener("click",()=>k("external")),n.addEventListener("click",I=>{I.target===n&&k(null)}),v.appendChild(w),v.appendChild(b),r.appendChild(s),r.appendChild(o),r.appendChild(l),r.appendChild(v),n.appendChild(r),document.body.appendChild(n)}),R="data-epnc-single-file-share";let me=0;const Ce=e=>{if(!e)return null;const n='a[href*="/s/'+e+'/download"],a[href*="/index.php/s/'+e+'/download"]',r=document.querySelector(n);return r instanceof HTMLAnchorElement?r:null},ct=()=>{const e=document.getElementById("initial-state-files_sharing-view");if(!(e instanceof HTMLInputElement))return"";const n=String(e.value||"").trim();if(n==="")return"";try{const r=window.atob(n);if(r.startsWith('"')||r.startsWith("{")||r.startsWith("[")){const o=JSON.parse(r);return typeof o=="string"?o:""}return r}catch{return""}},Ee=e=>{if(!e)return!1;const n=Ce(e);if(!n)return!1;let r;try{r=new URL(n.getAttribute("href")||"",window.location.origin)}catch{return!1}const o=(r.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!o||o[1]!==e?!1:!r.searchParams.has("files")},dt=()=>{const e=document.body;if(!(e instanceof HTMLElement))return{retry:!0};const n=$();if(!n)return e.removeAttribute(R),{retry:!1};const r=ct();return r==="public-file-share"?(e.setAttribute(R,"1"),{retry:!1}):r==="public-share"?(e.removeAttribute(R),{retry:!1}):Ce(n)?Ee(n)?(e.setAttribute(R,"1"),{retry:!1}):(e.removeAttribute(R),{retry:!1}):(e.removeAttribute(R),{retry:!0})},ut=()=>{const e=++me,n=r=>{e===me&&dt().retry&&r<30&&window.setTimeout(()=>n(r+1),180)};n(0)};(function(){const e="etherpad_nextcloud_open";let n=!1,r=null,o=0,s="";const l=Xe(),v=i=>{const a=Number(i);return Number.isFinite(a)&&a>0?a:null},w=i=>{if(!i)return null;const a=[i.fileId,i.fileid,i.id,i.nodeId,i.nodeid,i.file_id];for(const p of a){const y=v(p);if(y!==null)return y}if(typeof i.get=="function")for(const p of["fileid","fileId","id"])try{const y=v(i.get(p));if(y!==null)return y}catch{}return null},b=i=>{if(!(i instanceof Element))return null;const a=[i.getAttribute("data-fileid"),i.getAttribute("data-file-id"),i.getAttribute("data-id"),i.getAttribute("data-node-id")];for(const y of a){const m=v(y);if(m!==null)return m}const p=i.matches("a[href]")?i:i.querySelector("a[href]");return p instanceof HTMLAnchorElement?we(p.getAttribute("href")):null},C=i=>{if(!i||typeof i!="object")return null;const a=[i,i.fileInfoModel,i.fileInfo,i.model,i.node,i.file,i.attributes];for(const y of a){const m=w(y);if(m!==null)return m}const p=[i.fileElement,i.el,i.element,i.$file&&i.$file[0],i.$el&&i.$el[0],i.target,i.currentTarget];for(const y of p){const m=b(y);if(m!==null)return m}return null},k=async()=>{const i=await at();if(!i)return;const a=i.trim(),p=a.toLowerCase().endsWith(".pad")?a:a+".pad",y=B(Q(),p);try{const m=await Fe(y),T=m&&typeof m.file=="string"?m.file:y,E=m&&Number.isFinite(Number(m.file_id))?Number(m.file_id):null;await oe({path:T,fileId:E},{fallbackOpen:u,resolveOpenDir:X})}catch(m){const T=m instanceof Error?m.message:t(f,"Could not create public pad.");window.alert(T)}},I=async()=>{const i=await st();if(!i)return;const a=i.padUrl.trim(),p=i.name.trim(),y=Q(),m=p.toLowerCase().endsWith(".pad")?p:p+".pad",T=B(y,m);try{const E=await Pe(T,a),O=E&&typeof E.file=="string"?E.file:T,L=E&&Number.isFinite(Number(E.file_id))?Number(E.file_id):null;await oe({path:O,fileId:L},{fallbackOpen:u,resolveOpenDir:X})}catch(E){const O=E instanceof Error?E.message:t(f,"Could not import public pad URL.");window.alert(O)}},h=async()=>{const i=await lt();if(i==="internal"){await k();return}i==="external"&&await I()},d=(i,a)=>{const p=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return!p||typeof p.goToRoute!="function"||!i||!Number.isFinite(i)?!1:(p.goToRoute(null,{view:"files",fileid:String(i)},{dir:X(a),editing:"false",openfile:"true"}),window.setTimeout(()=>{if((window.location.pathname||"").includes("/apps/files/files/"+String(i)))return;const y=J(i,a);window.location.assign(y)},180),!0)},u=async i=>{const a=String(i.fileId??"")+"|"+String(i.path??""),p=Date.now();if(r===a&&p-o<800)return;r=a,o=p;const y=$(),m=y!==null&&y!=="",T=()=>{if(m){window.location.assign(G(y,i.path||""));return}if(F()&&i.fileId!==null&&i.fileId!==void 0&&Number.isFinite(Number(i.fileId))){const z=i.path||"/";window.location.assign(J(Number(i.fileId),z));return}const L=F()?j():null;if((i.fileId===null||i.fileId===void 0)&&L){const z=i.path||"/";window.location.assign(J(L,z));return}if(i.fileId!==null&&i.fileId!==void 0&&Number.isFinite(Number(i.fileId))){const z=i.path||"/";window.location.assign(J(Number(i.fileId),z));return}i.path&&window.location.assign(Me(i.path))};if(!H()){T();return}let E=i.path||"",O=i.fileId??null;if(!E&&i.fileId!==null&&i.fileId!==void 0)try{const L=await Z(i.fileId);E=L&&typeof L.path=="string"?L.path:"",O=L&&Number.isFinite(Number(L.file_id))?Number(L.file_id):O}catch{E=""}if(!E){T();return}if((!O||!Number.isFinite(Number(O)))&&E&&!m)try{const L=await ke(E);O=L&&Number.isFinite(Number(L.file_id))?Number(L.file_id):O}catch{}if((!O||!Number.isFinite(Number(O)))&&F()){const L=j();L&&(O=L)}if(F()){if(O&&Number.isFinite(Number(O))&&d(Number(O),E))return;T();return}try{window.OCA.Viewer.open({path:E})}catch{T()}},c=(i,a)=>{if(!(i instanceof HTMLAnchorElement)||!a)return null;const p=i.getAttribute("href")||"",y=Oe(p);if(y&&y.token===a&&K(y.path))return y.path;const m=Te(p);return K(m)?m:null},g=()=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const i=async a=>{if(!a||a.defaultPrevented)return;const p=$();if(!p||a.type==="click"&&(a.button!==0||a.metaKey||a.ctrlKey||a.shiftKey||a.altKey)||!(a.target instanceof Element))return;const y=a.target.closest("a[href]");if(!(y instanceof HTMLAnchorElement))return;const m=c(y,p);if(m){if(a.preventDefault(),a.stopPropagation(),typeof a.stopImmediatePropagation=="function"&&a.stopImmediatePropagation(),H()){await u({path:m,fileId:null});return}window.location.assign(G(p,m))}};document.addEventListener("click",i,!0)},A=()=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(ie,e,Se(),re(f,"etherpad-icon-color"),(i,a)=>{const p=a&&a.dir||Q(),y=B(p,i),m=C(a);u({path:y,fileId:m})},t(f,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(ie,e))},x=it({isFilesAppRoute:F,onCreatePublicPad:h}),S=async()=>{if(H())return;const i=j();if(!i)return;const a=new URLSearchParams(window.location.search||"");if(!(a.get("epNoRedirect")==="1"||a.get("details")==="1"||a.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+f+"/"))try{const p=await Z(i);p&&p.is_pad&&p.viewer_url&&window.location.assign(p.viewer_url)}catch{}},N=async()=>{if(H()||!F())return;const i=j();if(!i)return;const a=new URLSearchParams(window.location.search||"");if(!(a.get("openfile")==="true"||a.get("opendetails")==="true"||a.get("details")==="1"))try{const p=await Z(i);if(!p||!p.is_pad)return;const y=a.get("dir")||"/",m=_("/apps/files/files")+"?dir="+encodeURIComponent(y);window.location.href!==m&&window.location.assign(m)}catch{}},P=()=>{const i=$();if(!i)return;const a=new URLSearchParams(window.location.search||""),p=a.get("files")||"";if(H()){if(!K(p))return;const T=a.get("path")||"/",E=B(T,p);u({path:E,fileId:null});return}if(!K(p)){if((window.location.pathname||"").includes("/apps/"+f+"/public/"))return;Ee(i)&&window.location.assign(G(i,""));return}const y=a.get("path")||"/",m=B(y,p);window.location.assign(G(i,m))},M=()=>{const i=(window.location.pathname||"")+"?"+(window.location.search||"");i!==s&&(s=i,F()&&x(),ut(),N(),S(),P(),l.scheduleRefresh(0))},V=()=>{const i="EtherpadNextcloudRouteWatchInstalled";if(window.OCA&&window.OCA[i]===!0)return;window.OCA&&(window.OCA[i]=!0);const a=()=>window.setTimeout(M,0);window.addEventListener("popstate",a),window.addEventListener("hashchange",a);const p=y=>{const m=window.history&&window.history[y];typeof m=="function"&&(window.history[y]=function(...T){const E=m.apply(this,T);return a(),E})};p("pushState"),p("replaceState")},W=()=>{n||(n=!0,V(),M(),A(),g(),l.installObserver(),l.scheduleRefresh(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",W,{once:!0}):W()})();
+import{a as T,A as p,o as de,b as ge,c as ve,d as X,n as Ce,e as Ee,M as Y,f as xe}from"./oc-compat-Cn8-bqPr.chunk.mjs";import{p as U,a as z,v as D,f as j,b as Ae,r as Z,c as ue,g as K,n as R,i as B,d as Ne,e as Le}from"./urls-D-89WmZ3.chunk.mjs";const P=new Map,$=async n=>{const i=String(n);if(P.has(i))return P.get(i);const e=T("/apps/"+p+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(i),r=_(e,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve failed.").catch(a=>{throw P.delete(i),a});return P.set(i,r),r},Se=async n=>{const i="path:"+String(n);if(P.has(i))return P.get(i);const e=T("/apps/"+p+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(n),r=_(e,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve by path failed.").catch(a=>{throw P.delete(i),a});return P.set(i,r),r},V=async n=>{const i=T("/apps/"+p+"/api/v1/pads/sync-status/"+encodeURIComponent(String(n)));return _(i,{method:"GET",headers:{Accept:"application/json"}},"Sync status check failed.")},Oe=async(n,i)=>{let e=T("/apps/"+p+"/api/v1/pads/sync/"+encodeURIComponent(String(n)));return e+="?force=1",_(e,{method:"POST",headers:{Accept:"application/json",requesttoken:de()}},"Pad sync failed.")},Ie=async(n,i)=>{const e=T("/apps/"+p+"/api/v1/pads/from-url"),r=new URLSearchParams;return r.set("file",n),r.set("padUrl",i),_(e,{method:"POST",headers:pe(),body:r.toString()},"Could not import public pad URL.")},Me=async n=>{const i=T("/apps/"+p+"/api/v1/pads"),e=new URLSearchParams;return e.set("file",n),e.set("accessMode","public"),_(i,{method:"POST",headers:pe(),body:e.toString()},"Could not create public pad.")},pe=()=>({Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:de()}),_=async(n,i,e)=>{const r=await fetch(n,{...i,credentials:"same-origin"}),a=await r.json().catch(()=>({}));if(!r.ok)throw new Error(a&&a.message||e);return a},q=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),k=()=>(window.location.pathname||"").includes("/apps/files"),ke=120,Pe=900,ee=async(n,i={})=>{if(q()&&n.path)try{const e=Number(n.fileId);k()&&Number.isFinite(e)&&e>0&&(_e(e,n.path,i.resolveOpenDir),await Te()),await Fe(n.path),window.OCA.Viewer.open({path:n.path,onClose:He});return}catch{}typeof i.fallbackOpen=="function"&&await i.fallbackOpen(n)},Te=()=>new Promise(n=>{window.setTimeout(n,ke)}),Fe=async n=>{const i=ge(n);i!==""&&(ve("editor:file:created",i),await qe())},qe=()=>new Promise(n=>{window.setTimeout(n,Pe)}),_e=(n,i,e)=>{const r=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;!r||typeof r.goToRoute!="function"||r.goToRoute(null,{...r.params||{},view:r.params&&r.params.view?r.params.view:"files",fileid:String(n)},{...r.query||{},dir:Re(i,e),editing:"false",openfile:"true"},!0)},Re=(n,i)=>{if(typeof i=="function")return i(n);const e=String(n||""),r=e.lastIndexOf("/");return r>0?e.substring(0,r):"/"},He=()=>{const n=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;if(!n||typeof n.goToRoute!="function")return;const i={...n.query||{}};delete i.openfile,delete i.editing,n.goToRoute(null,n.params||{},i)},Ue=800,ze=180,Be=(n,i)=>{const e=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return!e||typeof e.goToRoute!="function"||!n||!Number.isFinite(n)?!1:(e.goToRoute(null,{view:"files",fileid:String(n)},{dir:Z(i),editing:"false",openfile:"true"}),window.setTimeout(()=>{if((window.location.pathname||"").includes("/apps/files/files/"+String(n)))return;const r=j(n,i);window.location.assign(r)},ze),!0)},De=()=>{let n=null,i=0;return async e=>{const r=String(e.fileId??"")+"|"+String(e.path??""),a=Date.now();if(n===r&&a-i<Ue)return;n=r,i=a;const s=U(),E=s!==null&&s!=="",f=()=>{if(E){window.location.assign(D(s,e.path||""));return}if(k()&&e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const N=e.path||"/";window.location.assign(j(Number(e.fileId),N));return}const g=k()?z():null;if((e.fileId===null||e.fileId===void 0)&&g){const N=e.path||"/";window.location.assign(j(g,N));return}if(e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const N=e.path||"/";window.location.assign(j(Number(e.fileId),N));return}e.path&&window.location.assign(Ae(e.path))};if(!q()){f();return}let y=e.path||"",w=e.fileId??null;if(!y&&e.fileId!==null&&e.fileId!==void 0)try{const g=await $(e.fileId);y=g&&typeof g.path=="string"?g.path:"",w=g&&Number.isFinite(Number(g.file_id))?Number(g.file_id):w}catch{y=""}if(!y){f();return}if((!w||!Number.isFinite(Number(w)))&&y&&!E)try{const g=await Se(y);w=g&&Number.isFinite(Number(g.file_id))?Number(g.file_id):w}catch{}if((!w||!Number.isFinite(Number(w)))&&k()){const g=z();g&&(w=g)}if(k()){if(w&&Number.isFinite(Number(w))&&Be(Number(w),y))return;f();return}try{window.OCA.Viewer.open({path:y})}catch{f()}}},G="data-epnc-sidebar-sync-panel",Q="data-epnc-sidebar-sync-mount",fe=8e3,je=6e4,me=15e3,$e=8e3,We=()=>{const n=document.documentElement,i=document.body,e=(n?.className||"")+" "+(i?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(e))return!0;const r=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(r&&r.matches)},te=n=>{if(!(n instanceof HTMLElement))return!1;const i=window.getComputedStyle(n);return i.display==="none"||i.visibility==="hidden"?!1:i.position==="fixed"?!0:n.offsetParent!==null},Ke=n=>{const i=Number(n);return Number.isFinite(i)&&i>0?i:null},Ve=()=>{const n=z();if(n!==null)return n;const i=new URLSearchParams(window.location.search||""),e=Number(i.get("fileid")||"");return Number.isFinite(e)&&e>0?e:null},ne=()=>{const n=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const i of n){const e=document.querySelector(i);if(e instanceof HTMLElement)return e}return null},Ge=n=>{if(!(n instanceof HTMLElement))return null;let i=n.querySelector("["+Q+'="1"]');i instanceof HTMLElement||(i=document.createElement("div"),i.className="epnc-sidebar-sync-mount",i.setAttribute(Q,"1"));const e=n.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(e&&e.parentNode instanceof HTMLElement){const r=e.parentNode;return(i.parentNode!==r||i.nextSibling!==e)&&r.insertBefore(i,e),i}return i.parentNode!==n&&n.insertBefore(i,n.firstChild),i},ie=n=>{const i=Ve();if(i!==null)return i;if(!(n instanceof HTMLElement))return null;const e=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const a of e){const s=n.querySelector(a);if(!(s instanceof HTMLElement))continue;const E=[s.getAttribute("data-fileid"),s.getAttribute("data-file-id"),s.getAttribute("data-node-id"),s.getAttribute("data-id"),s.dataset&&(s.dataset.fileid||s.dataset.fileId)];for(const f of E){const y=Ke(f);if(y!==null)return y}}const r=n.querySelectorAll("a[href]");for(const a of r){if(!(a instanceof HTMLAnchorElement))continue;const s=ue(a.getAttribute("href"));if(s!==null)return s}return null},Je=n=>n==="synced"?"etherpad-icon-color":We()?"etherpad-icon-white":"etherpad-icon-black",Ze=n=>n==="synced"?t(p,"Synchronized"):n==="syncing"?t(p,"Saving..."):n==="error"?t(p,"Sync failed"):n==="unavailable"?t(p,"Sync status unavailable"):n==="loading"?t(p,"Checking sync status..."):t(p,"Sync pending"),J=n=>n==="synced"?"synced":n==="out_of_sync"?"pending":n==="unavailable"?"unavailable":"error",Qe=(n,i)=>n==="synced"?Math.min(je,i+$e):n==="pending"?fe:me,Xe=()=>{let n=null,i=null,e=null,r=0;const a=()=>{e!==null&&(window.clearTimeout(e),e=null)},s=()=>{const d=document.querySelector("["+G+'="1"]');d&&d.parentNode&&d.parentNode.removeChild(d);const u=document.querySelector("["+Q+'="1"]');u&&u.parentNode&&u.parentNode.removeChild(u),a()},E=(d,u)=>{if(!(d instanceof HTMLElement))return;const c=String(u||"").trim();d.setAttribute("data-public-open-url",c);const C=d.querySelector("[data-epnc-sidebar-open-public-button]");if(!(C instanceof HTMLButtonElement))return;const A=c!=="";C.style.display=A?"inline-flex":"none",C.disabled=!A},f=(d,u,c,C,A=!0)=>{if(!(d instanceof HTMLElement))return;const x=d.querySelector("[data-epnc-sidebar-sync-icon]"),S=d.querySelector("[data-epnc-sidebar-sync-label]"),L=d.querySelector("[data-epnc-sidebar-sync-button]");if(d.setAttribute("data-file-id",String(c)),d.setAttribute("data-sync-state",u),d.setAttribute("data-sync-in-flight",C?"1":"0"),S instanceof HTMLElement&&(S.textContent=Ze(u)),x instanceof HTMLElement){const I=Je(u),o=X(p,I);o!==""&&(x.style.backgroundImage=`url(${o})`)}L instanceof HTMLButtonElement&&(L.disabled=C||!A||c<=0,L.setAttribute("aria-busy",C?"true":"false"))},y=(d,u)=>{if(!(u instanceof HTMLElement))return null;let c=u.querySelector("["+G+'="1"]');if(!(c instanceof HTMLElement)){c=document.createElement("section"),c.className="epnc-sidebar-sync-panel",c.setAttribute(G,"1"),c.setAttribute("data-file-id",String(d));const C=document.createElement("div");C.className="epnc-sidebar-sync-panel__header";const A=document.createElement("span");A.className="epnc-sidebar-sync-panel__icon",A.setAttribute("data-epnc-sidebar-sync-icon","1"),A.setAttribute("aria-hidden","true");const x=document.createElement("span");x.className="epnc-sidebar-sync-panel__status",x.setAttribute("data-epnc-sidebar-sync-label","1"),C.appendChild(A),C.appendChild(x);const S=document.createElement("button");S.type="button",S.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",S.setAttribute("data-epnc-sidebar-sync-button","1"),S.textContent=t(p,"Pad in Datei speichern"),S.addEventListener("click",async I=>{I.preventDefault(),I.stopPropagation();const o=Number(c.getAttribute("data-file-id")||"");if(!(!Number.isFinite(o)||o<=0)){f(c,"syncing",o,!0,!0);try{await Oe(o,!0);const l=await V(o),m=J(l&&l.status);f(c,m,o,!1,!0)}catch{f(c,"error",o,!1,!0)}}});const L=document.createElement("button");L.type="button",L.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",L.setAttribute("data-epnc-sidebar-open-public-button","1"),L.textContent=t(p,"Pad in neuem Tab öffnen"),L.style.display="none",L.addEventListener("click",I=>{I.preventDefault(),I.stopPropagation();const o=String(c.getAttribute("data-public-open-url")||"").trim();o!==""&&window.open(o,"_blank","noopener,noreferrer")}),c.appendChild(C),c.appendChild(S),c.appendChild(L)}return c.parentNode!==u&&u.appendChild(c),c},w=(d,u)=>{if(a(),!(d instanceof HTMLElement)||!Number.isFinite(u)||u<=0)return;let c=fe;const C=x=>{e=window.setTimeout(()=>{A()},Math.max(1e3,x))},A=async()=>{if(!(d instanceof HTMLElement)||!document.body.contains(d)){a();return}const x=ne();if(!(x instanceof HTMLElement)||!te(x)||!te(d)){a();return}if(ie(x)!==u){a();return}if(document.visibilityState!=="visible"||d.getAttribute("data-sync-in-flight")==="1"){C(c);return}try{const S=await V(u),L=J(S&&S.status);f(d,L,u,!1,!0),c=Qe(L,c)}catch{f(d,"error",u,!1,!0),c=me}C(c)};C(c)},g=()=>{i!==null&&(window.clearTimeout(i),i=null)},N=async()=>{const d=++r;if(!k()||U()){s();return}const u=ne();if(!(u instanceof HTMLElement)){s();return}const c=ie(u);if(!c){s();return}const C=Ge(u);if(!(C instanceof HTMLElement)){s();return}let A;try{A=await $(c)}catch{s();return}if(d!==r)return;if(!A||A.is_pad!==!0){s();return}const x=y(c,C);if(!(x instanceof HTMLElement)){s();return}const S=A&&typeof A.public_open_url=="string"?A.public_open_url.trim():"";E(x,S),f(x,"loading",c,!1,!0);try{const L=await V(c);if(d!==r)return;const I=J(L&&L.status);f(x,I,c,!1,!0),w(x,c)}catch{if(d!==r)return;f(x,"error",c,!1,!0),w(x,c)}},v=(d=0)=>{g(),i=window.setTimeout(()=>{i=null,N()},Math.max(0,d))};return{installObserver:()=>{if(n!==null||!k())return;const d=new MutationObserver(u=>{for(const c of u)for(const C of c.addedNodes)if(C instanceof Element&&(C.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||C.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){v(60);return}});d.observe(document.body,{childList:!0,subtree:!0}),n=d},scheduleRefresh:v}},re=p+"_public_pad",oe="icon-filetype-etherpad-nextcloud-pad",ae='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',se=98,Ye=120,et=500,le=()=>{const n=Ce();return n!==null&&n>=33},H=n=>(n&&n.message?String(n.message):"").toLowerCase().includes("duplicate"),tt=n=>{if(!n||typeof n!="object")return null;if(n.hasCreatePermission===!1)return!1;if(n.hasCreatePermission===!0)return!0;if(typeof n.canCreate=="boolean")return n.canCreate;const i=Ee(),e=[n.permissions,n.permission,n.attributes&&n.attributes.permissions,typeof n.get=="function"?n.get("permissions"):null];for(const r of e){const a=Number(r);if(Number.isFinite(a))return(a&i)===i}return null},nt=(...n)=>{for(const i of n){const e=tt(i);if(e!==null)return e}return!0},it=()=>{const n=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const e of n)if(!(!e||typeof e!="object")){if(typeof e.addNewFileMenuEntry=="function")return{register:e.addNewFileMenuEntry.bind(e),unregister:typeof e.removeNewFileMenuEntry=="function"?e.removeNewFileMenuEntry.bind(e):null};if(typeof e.getNewFileMenu=="function"){let r=null;try{r=e.getNewFileMenu()}catch{r=null}if(r&&typeof r.registerEntry=="function")return{register:r.registerEntry.bind(r),unregister:typeof r.unregisterEntry=="function"?r.unregisterEntry.bind(r):null}}}const i=window._nc_newfilemenu;return i&&typeof i=="object"&&typeof i.registerEntry=="function"?{register:i.registerEntry.bind(i),unregister:typeof i.unregisterEntry=="function"?i.unregisterEntry.bind(i):null}:null},rt=({isFilesAppRoute:n,onCreatePublicPad:i})=>{let e=!1,r=!1,a=!1,s=0;const E=()=>{i()},f=(...v)=>nt(...v),y=()=>({id:re,displayName:t(p,"Public pad"),...le()?{iconSvgInline:ae}:{iconClass:oe},order:se,enabled:f,handler:E}),w=()=>({id:re,displayName:t(p,"Public pad"),templateName:t(p,"Public pad"),iconClass:oe,...le()?{iconSvgInline:ae}:{},fileType:"file",order:se,actionHandler:E}),g=async()=>{const v=it();if(!v)return!1;const d=y();try{return v.register(d),e=!0,!0}catch(u){return H(u)?(e=!0,!0):!1}},N=()=>{if(r)return!0;const v=w(),d=()=>{const u=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const c of u)if(!(!c||typeof c!="object")){if(typeof c.addMenuEntry=="function")try{return c.addMenuEntry(v),!0}catch(C){if(H(C))return!0}if(typeof c.registerEntry=="function")try{return c.registerEntry(y()),!0}catch(C){if(H(C))return!0}}return!1};try{return d()?(r=!0,!0):(!a&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(u){if(!(!u||typeof u.addMenuEntry!="function"))try{u.addMenuEntry(v),r=!0}catch(c){H(c)&&(r=!0)}}}),a=!0),r)}catch(u){return H(u)?(r=!0,!0):!1}};return()=>{if(!n()||e||r)return;const v=++s,d=async u=>{v!==s||e||r||await g()||N()||u<Ye&&window.setTimeout(()=>{d(u+1)},et)};d(0)}},ot=n=>{try{const i=new URL(n,window.location.origin),e=decodeURIComponent(i.pathname||"").match(/\/p\/([^/]+)$/);if(e&&e[1]){const r=e[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(r!=="")return r+".pad"}}catch{}return"Imported pad.pad"},ye=n=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const e=document.createElement("div");e.style.position="relative",e.style.background="var(--color-main-background, #fff)",e.style.borderRadius="10px",e.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",e.style.padding="18px",e.style.width="min(460px, calc(100vw - 24px))";const r=document.createElement("button");r.type="button",r.setAttribute("aria-label",t(p,"Close")),r.textContent="×",r.style.position="absolute",r.style.top="8px",r.style.right="10px",r.style.border="none",r.style.background="transparent",r.style.fontSize="22px",r.style.cursor="pointer",r.style.lineHeight="1";const a=document.createElement("h3");return a.textContent=n,a.style.margin="0 26px 10px 0",a.style.fontSize="18px",e.appendChild(r),e.appendChild(a),i.appendChild(e),document.body.appendChild(i),{overlay:i,dialog:e,closeButton:r}},at=()=>new Promise(n=>{const{overlay:i,dialog:e,closeButton:r}=ye(t(p,"Internal public pad")),a=document.createElement("label");a.textContent=t(p,"File name"),a.style.display="block",a.style.marginBottom="6px";const s=document.createElement("input");s.type="text",s.value=t(p,"Public pad")+".pad",s.style.width="100%",s.style.boxSizing="border-box",s.style.marginBottom="12px";const E=document.createElement("p");E.style.color="var(--color-error, #c62828)",E.style.margin="0 0 12px 0",E.style.minHeight="20px";const f=document.createElement("button");f.type="button",f.className="primary",f.textContent=t(p,"Create");const y=w=>{i.remove(),n(w)};r.addEventListener("click",()=>y(null)),i.addEventListener("click",w=>{w.target===i&&y(null)}),s.addEventListener("keydown",w=>{w.key==="Enter"&&(w.preventDefault(),f.click())}),f.addEventListener("click",()=>{const w=s.value.trim();if(w===""){E.textContent=t(p,"File name is required.");return}y(w)}),e.appendChild(a),e.appendChild(s),e.appendChild(E),e.appendChild(f),s.focus(),s.select()}),st=()=>new Promise(n=>{const{overlay:i,dialog:e,closeButton:r}=ye(t(p,"External public pad (URL)")),a=document.createElement("label");a.textContent=t(p,"External pad URL"),a.style.display="block",a.style.marginBottom="6px";const s=document.createElement("input");s.type="url",s.value="https://",s.placeholder="https://",s.style.width="100%",s.style.boxSizing="border-box",s.style.marginBottom="12px";const E=document.createElement("label");E.textContent=t(p,"File name"),E.style.display="block",E.style.marginBottom="6px";const f=document.createElement("input");f.type="text",f.value="Imported pad.pad",f.style.width="100%",f.style.boxSizing="border-box",f.style.marginBottom="12px";const y=document.createElement("p");y.style.color="var(--color-error, #c62828)",y.style.margin="0 0 12px 0",y.style.minHeight="20px";const w=document.createElement("button");w.type="button",w.className="primary",w.textContent=t(p,"Create");const g=v=>{i.remove(),n(v)};r.addEventListener("click",()=>g(null)),i.addEventListener("click",v=>{v.target===i&&g(null)}),s.addEventListener("blur",()=>{const v=s.value.trim();v.startsWith("http")&&(f.value=ot(v))});const N=()=>{const v=s.value.trim(),d=f.value.trim();if(v===""){y.textContent=t(p,"External pad URL is required.");return}if(d===""){y.textContent=t(p,"File name is required.");return}g({padUrl:v,name:d})};s.addEventListener("keydown",v=>{v.key==="Enter"&&(v.preventDefault(),N())}),f.addEventListener("keydown",v=>{v.key==="Enter"&&(v.preventDefault(),N())}),w.addEventListener("click",N),e.appendChild(a),e.appendChild(s),e.appendChild(E),e.appendChild(f),e.appendChild(y),e.appendChild(w),s.focus(),s.select()}),lt=()=>new Promise(n=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const e=document.createElement("div");e.style.position="relative",e.style.background="var(--color-main-background, #fff)",e.style.borderRadius="10px",e.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",e.style.padding="18px",e.style.width="min(420px, calc(100vw - 24px))";const r=document.createElement("h3");r.textContent=t(p,"Create public pad"),r.style.margin="0 26px 10px 0",r.style.fontSize="18px";const a=document.createElement("button");a.type="button",a.setAttribute("aria-label",t(p,"Close")),a.textContent="×",a.style.position="absolute",a.style.top="8px",a.style.right="10px",a.style.border="none",a.style.background="transparent",a.style.fontSize="22px",a.style.cursor="pointer",a.style.lineHeight="1";const s=document.createElement("p");s.textContent=t(p,"Choose how you want to create the public pad."),s.style.margin="0 0 14px 0",s.style.opacity="0.8";const E=document.createElement("div");E.style.display="grid",E.style.gap="8px";const f=document.createElement("button");f.type="button",f.className="primary",f.textContent=t(p,"Internal public pad");const y=document.createElement("button");y.type="button",y.textContent=t(p,"External public pad (URL)");const w=X(p,"etherpad-icon-color");w!==""&&(y.style.backgroundImage=`url(${w})`),y.style.backgroundRepeat="no-repeat",y.style.backgroundPosition="12px center",y.style.backgroundSize="16px 16px",y.style.paddingLeft="34px";const g=N=>{i.remove(),n(N)};a.addEventListener("click",()=>g(null)),f.addEventListener("click",()=>g("internal")),y.addEventListener("click",()=>g("external")),i.addEventListener("click",N=>{N.target===i&&g(null)}),E.appendChild(f),E.appendChild(y),e.appendChild(a),e.appendChild(r),e.appendChild(s),e.appendChild(E),i.appendChild(e),document.body.appendChild(i)}),F="data-epnc-single-file-share";let ce=0;const we=n=>{if(!n)return null;const i='a[href*="/s/'+n+'/download"],a[href*="/index.php/s/'+n+'/download"]',e=document.querySelector(i);return e instanceof HTMLAnchorElement?e:null},ct=()=>{const n=document.getElementById("initial-state-files_sharing-view");if(!(n instanceof HTMLInputElement))return"";const i=String(n.value||"").trim();if(i==="")return"";try{const e=window.atob(i);if(e.startsWith('"')||e.startsWith("{")||e.startsWith("[")){const r=JSON.parse(e);return typeof r=="string"?r:""}return e}catch{return""}},be=n=>{if(!n)return!1;const i=we(n);if(!i)return!1;let e;try{e=new URL(i.getAttribute("href")||"",window.location.origin)}catch{return!1}const r=(e.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!r||r[1]!==n?!1:!e.searchParams.has("files")},dt=()=>{const n=document.body;if(!(n instanceof HTMLElement))return{retry:!0};const i=U();if(!i)return n.removeAttribute(F),{retry:!1};const e=ct();return e==="public-file-share"?(n.setAttribute(F,"1"),{retry:!1}):e==="public-share"?(n.removeAttribute(F),{retry:!1}):we(i)?be(i)?(n.setAttribute(F,"1"),{retry:!1}):(n.removeAttribute(F),{retry:!1}):(n.removeAttribute(F),{retry:!0})},ut=()=>{const n=++ce,i=e=>{n===ce&&dt().retry&&e<30&&window.setTimeout(()=>i(e+1),180)};i(0)};(function(){const n="etherpad_nextcloud_open";let i=!1,e="";const r=Xe(),a=De(),s=o=>{const l=Number(o);return Number.isFinite(l)&&l>0?l:null},E=o=>{if(!o)return null;const l=[o.fileId,o.fileid,o.id,o.nodeId,o.nodeid,o.file_id];for(const m of l){const h=s(m);if(h!==null)return h}if(typeof o.get=="function")for(const m of["fileid","fileId","id"])try{const h=s(o.get(m));if(h!==null)return h}catch{}return null},f=o=>{if(!(o instanceof Element))return null;const l=[o.getAttribute("data-fileid"),o.getAttribute("data-file-id"),o.getAttribute("data-id"),o.getAttribute("data-node-id")];for(const h of l){const b=s(h);if(b!==null)return b}const m=o.matches("a[href]")?o:o.querySelector("a[href]");return m instanceof HTMLAnchorElement?ue(m.getAttribute("href")):null},y=o=>{if(!o||typeof o!="object")return null;const l=[o,o.fileInfoModel,o.fileInfo,o.model,o.node,o.file,o.attributes];for(const h of l){const b=E(h);if(b!==null)return b}const m=[o.fileElement,o.el,o.element,o.$file&&o.$file[0],o.$el&&o.$el[0],o.target,o.currentTarget];for(const h of m){const b=f(h);if(b!==null)return b}return null},w=async()=>{const o=await at();if(!o)return;const l=o.trim(),m=l.toLowerCase().endsWith(".pad")?l:l+".pad",h=R(K(),m);try{const b=await Me(h),M=b&&typeof b.file=="string"?b.file:h,O=b&&Number.isFinite(Number(b.file_id))?Number(b.file_id):null;await ee({path:M,fileId:O},{fallbackOpen:a,resolveOpenDir:Z})}catch(b){const M=b instanceof Error?b.message:t(p,"Could not create public pad.");window.alert(M)}},g=async()=>{const o=await st();if(!o)return;const l=o.padUrl.trim(),m=o.name.trim(),h=K(),b=m.toLowerCase().endsWith(".pad")?m:m+".pad",M=R(h,b);try{const O=await Ie(M,l),W=O&&typeof O.file=="string"?O.file:M,he=O&&Number.isFinite(Number(O.file_id))?Number(O.file_id):null;await ee({path:W,fileId:he},{fallbackOpen:a,resolveOpenDir:Z})}catch(O){const W=O instanceof Error?O.message:t(p,"Could not import public pad URL.");window.alert(W)}},N=async()=>{const o=await lt();if(o==="internal"){await w();return}o==="external"&&await g()},v=(o,l)=>{if(!(o instanceof HTMLAnchorElement)||!l)return null;const m=o.getAttribute("href")||"",h=Ne(m);if(h&&h.token===l&&B(h.path))return h.path;const b=Le(m);return B(b)?b:null},d=()=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const o=async l=>{if(!l||l.defaultPrevented)return;const m=U();if(!m||l.type==="click"&&(l.button!==0||l.metaKey||l.ctrlKey||l.shiftKey||l.altKey)||!(l.target instanceof Element))return;const h=l.target.closest("a[href]");if(!(h instanceof HTMLAnchorElement))return;const b=v(h,m);if(b){if(l.preventDefault(),l.stopPropagation(),typeof l.stopImmediatePropagation=="function"&&l.stopImmediatePropagation(),q()){await a({path:b,fileId:null});return}window.location.assign(D(m,b))}};document.addEventListener("click",o,!0)},u=()=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(Y,n,xe(),X(p,"etherpad-icon-color"),(o,l)=>{const m=l&&l.dir||K(),h=R(m,o),b=y(l);a({path:h,fileId:b})},t(p,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(Y,n))},c=rt({isFilesAppRoute:k,onCreatePublicPad:N}),C=async()=>{if(q())return;const o=z();if(!o)return;const l=new URLSearchParams(window.location.search||"");if(!(l.get("epNoRedirect")==="1"||l.get("details")==="1"||l.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+p+"/"))try{const m=await $(o);m&&m.is_pad&&m.viewer_url&&window.location.assign(m.viewer_url)}catch{}},A=async()=>{if(q()||!k())return;const o=z();if(!o)return;const l=new URLSearchParams(window.location.search||"");if(!(l.get("openfile")==="true"||l.get("opendetails")==="true"||l.get("details")==="1"))try{const m=await $(o);if(!m||!m.is_pad)return;const h=l.get("dir")||"/",b=T("/apps/files/files")+"?dir="+encodeURIComponent(h);window.location.href!==b&&window.location.assign(b)}catch{}},x=()=>{const o=U();if(!o)return;const l=new URLSearchParams(window.location.search||""),m=l.get("files")||"";if(q()){if(!B(m))return;const M=l.get("path")||"/",O=R(M,m);a({path:O,fileId:null});return}if(!B(m)){if((window.location.pathname||"").includes("/apps/"+p+"/public/"))return;be(o)&&window.location.assign(D(o,""));return}const h=l.get("path")||"/",b=R(h,m);window.location.assign(D(o,b))},S=()=>{const o=(window.location.pathname||"")+"?"+(window.location.search||"");o!==e&&(e=o,k()&&c(),ut(),A(),C(),x(),r.scheduleRefresh(0))},L=()=>{const o="EtherpadNextcloudRouteWatchInstalled";if(window.OCA&&window.OCA[o]===!0)return;window.OCA&&(window.OCA[o]=!0);const l=()=>window.setTimeout(S,0);window.addEventListener("popstate",l),window.addEventListener("hashchange",l);const m=h=>{const b=window.history&&window.history[h];typeof b=="function"&&(window.history[h]=function(...M){const O=b.apply(this,M);return l(),O})};m("pushState"),m("replaceState")},I=()=>{i||(i=!0,L(),S(),u(),d(),r.installObserver(),r.scheduleRefresh(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",I,{once:!0}):I()})();
 //# sourceMappingURL=etherpad_nextcloud-files-main.mjs.map
